# internal/api/oas_schemas_gen.go
// Code generated by ogen, DO NOT EDIT.

package api

// Ref: #/components/schemas/Cat
type Cat struct {
	// The unique identifier of the cat.
	ID   int64  `json:"id"`
	Name string `json:"name"`
	// The age of the cat.
	Age int32 `json:"age"`
	// The color of the cat.
	Color string `json:"color"`
}

// GetID returns the value of ID.
func (s *Cat) GetID() int64 {
	return s.ID
}

// GetName returns the value of Name.
func (s *Cat) GetName() string {
	return s.Name
}

// GetAge returns the value of Age.
func (s *Cat) GetAge() int32 {
	return s.Age
}

// GetColor returns the value of Color.
func (s *Cat) GetColor() string {
	return s.Color
}

// SetID sets the value of ID.
func (s *Cat) SetID(val int64) {
	s.ID = val
}

// SetName sets the value of Name.
func (s *Cat) SetName(val string) {
	s.Name = val
}

// SetAge sets the value of Age.
func (s *Cat) SetAge(val int32) {
	s.Age = val
}

// SetColor sets the value of Color.
func (s *Cat) SetColor(val string) {
	s.Color = val
}

func (*Cat) getCatRes()    {}
func (*Cat) updateCatRes() {}

// DeleteCatNoContent is response for DeleteCat operation.
type DeleteCatNoContent struct{}

func (*DeleteCatNoContent) deleteCatRes() {}

// DeleteCatNotFound is response for DeleteCat operation.
type DeleteCatNotFound struct{}

func (*DeleteCatNotFound) deleteCatRes() {}

// GetCatNotFound is response for GetCat operation.
type GetCatNotFound struct{}

func (*GetCatNotFound) getCatRes() {}

// Parameters for creating a new cat.
// Ref: #/components/schemas/NewCat
type NewCat struct {
	Name  string `json:"name"`
	Age   int32  `json:"age"`
	Color string `json:"color"`
}

// GetName returns the value of Name.
func (s *NewCat) GetName() string {
	return s.Name
}

// GetAge returns the value of Age.
func (s *NewCat) GetAge() int32 {
	return s.Age
}

// GetColor returns the value of Color.
func (s *NewCat) GetColor() string {
	return s.Color
}

// SetName sets the value of Name.
func (s *NewCat) SetName(val string) {
	s.Name = val
}

// SetAge sets the value of Age.
func (s *NewCat) SetAge(val int32) {
	s.Age = val
}

// SetColor sets the value of Color.
func (s *NewCat) SetColor(val string) {
	s.Color = val
}

// Complete replacement (PUT) of an existing cat.
// Ref: #/components/schemas/UpdateCat
type UpdateCat struct {
	Name  string `json:"name"`
	Age   int32  `json:"age"`
	Color string `json:"color"`
}

// GetName returns the value of Name.
func (s *UpdateCat) GetName() string {
	return s.Name
}

// GetAge returns the value of Age.
func (s *UpdateCat) GetAge() int32 {
	return s.Age
}

// GetColor returns the value of Color.
func (s *UpdateCat) GetColor() string {
	return s.Color
}

// SetName sets the value of Name.
func (s *UpdateCat) SetName(val string) {
	s.Name = val
}

// SetAge sets the value of Age.
func (s *UpdateCat) SetAge(val int32) {
	s.Age = val
}

// SetColor sets the value of Color.
func (s *UpdateCat) SetColor(val string) {
	s.Color = val
}

// UpdateCatNotFound is response for UpdateCat operation.
type UpdateCatNotFound struct{}

func (*UpdateCatNotFound) updateCatRes() {}